vmm: move the http api into its own submodule
This commits moves the http API code into its own
submodule.
